requirements.txt
# Requirements for Stoncs hackathon demo
pandas>=1.5
requests>=2.28
# Note: Snowflake Python connector and Snowpark are intentionally omitted because
# this project uses Snowflake's REST API wrapper (`stoncs.snowflake_api_client`).
numpy>=1.23
scikit-learn>=1.1
sentence-transformers>=2.2.2
streamlit>=1.18
altair>=4.2
wordcloud>=1.8
matplotlib>=3.5
yfinance>=0.2
newsapi-python>=0.2
pytest>=7.0
joblib>=1.2
python-dotenv>=1.0
